Add demo and description for element template
Currently there is no demo reference in the element template.
Thus when creating a new element and serving it, there is no demo option, although the demo folder and demo template are created. It will be more consistent to include the demo refernce given that the demo template is provided.