small renaming and adding of logging to build/bundle
This is in preparation for the Windows fix I've just completed. I'd
like to push these seperately so that the windows fix can be as
concise as possible.

In this commit I added logging that helps detect the issue of missing
duplicate matches during bundling. I also renamed the path/osPath
variable names to be more intuitive (`path` now refers to the path
library).