Fix build bundling on windows
Previously, we were relying on path.posix.normalize() to convert unix &
windows paths to the posix path format, to be returned as a URL path.
However, path.posix.normalize() doesn't convert windows paths to posix.

Example:
```
> path.posix.normalize('/src/123.html')
'/src/123.html'
> path.posix.normalize('\\src\\123.html')
'\\src\\123.html // expected url format `/src/123/html`
```

Both Justin and I have looked into this, and the conclusion is that
Node does not support this conversion natively. This commit adds the
slash library to handle this conversion instead. This is a simple,
well-supported package used by babel and some other major projects.

Our original worry was that escaped back slashes would be incorrectly
converted, but Windows will actually not allow any slashes in file
names, escaped or otherwise. See:
http://stackoverflow.com/a/10708449
https://msdn.microsoft.com/en-us/library/windows/desktop/aa365247(v=vs.85).aspx

On Mac & Unix, backslashes are allowed in file names. This means that
slash will not properly convert those strings:

Example:
```
> path.posix.normalize('/src/this_file_has_escaped\\_see.html')
'/src/this_file_has_escaped/_see.html'
```

However, the files we are converting and checking against are all meant
for HTML import URL paths. Backward slashes are not allowed in URLs (RFC 2396)
and most browsers will just convert them to forward slashes. So this
case does not need to be supported. One possible enhancement would be
to throw a helpful error in this case, because right now we'll just
silently convert it to a forward slash seperator.