Accept relative roots in project config.
Most file paths in the project config are resolved to their absolute
paths by the time they are passed to `build`. For some reason, this is
not true for `root`. This change resolves the configured `root` to its
absolute path, so that
[a later check](https://github.com/Polymer/polymer-cli/blob/master/src/build/url-from-path.ts#L43-L45)
comparing root to other file paths will succeed when appropriate.