Fix sub-generator names at init-prompt
The generator name was parsed from the directory name
(./generator-polymer-init-foo/ --> "foo"), and this name was
given to all sub-generators in the directory, regardless of their
individual package.json config for 'name', which made them
indistinguishable without descriptions. Technically, Yeoman
generators are supposed to be named based on their package.json
config.

This patch properly parses the generator name from package.json,
and falls back to the name parsed from the directory name if
necessary.