Fix an issue when compiling with the latest TSC.
The latest version of typescript's standard library requires that the keys of a WeakMap be objects. So

interface WeakMap<K,V> { ... }

became

interface WeakMap<K extends object, V> { ... }

This conflicts with a declaration of WeakMap in the lodash typings (which the CLI depends on transitively).

Temporary workaround is to not typecheck libs.

More info: https://github.com/DefinitelyTyped/DefinitelyTyped/issues/14324