Add .polymerignore mechanism
Several templates for the init command are stored on github:
- polymer-1-starter-kit
- polymer-2-starter-kit
- shop

These projects have several files that are not meant to be used by
consumers of the template.  Examples:
- .travis.yml
- .github
- .gitattributes

The solution offered in this commit is provides a mechanism to read a
`.polymerignore` file that filters for the purpose of `polymer init`
templates.  This way, any new files that get added to the template do
not need to be registered in polymer-cli.  The `.polymerignore` format
is very much like the `.gitignore` format.
- Each line is a glob pattern (both * and ** are respected).
- Empty lines are ignored.
- Lines that begin with # are ignored.
- Lines that begin with / are considered to be paths at the root of the
  project.

Issues:
- #446
- #629