Exit with explicit status 0 instead of unspecified
When a node process is just interrupted via a SIGINT, it will report an undefined exit code. [The FAQ of Concurrently](https://www.npmjs.com/package/concurrently#faq) explicitly states that in such a case, it will exit with a non-null status, causing npm's runner to exit with an error. This change fixes that issue.