Reduce the amount of noise in test output.
This captures all non-plylog output and lets us assert on it rather than logging it from the test runner.

It does not intercept plylog, as plylog is really hard to intercept. For the editor service, when it was a hard requirement, the only way I was able to do it was by importing a special log interceptor module as the very first thing.